refactor: i dont even know what i did
